Update index.rst
In the section "Two-way data sync," the documentation mentions that data can be synced to all devices from the server. However, it is not entirely clear whether this is an automatic process or if users need to manually trigger the data sync. Providing clarification on this point would help users understand the functionality of the tool better.