Update index.rst
Upon further review, the previous statement indicating that data can be synced to all devices from the server either automatically or manually triggered by the user is incorrect. The synchronization process requires manual input from the user, as they need to click the sync button for the data to be synced between the server and the device. I apologize for any confusion this may have caused and aim to provide accurate information to our users.